# pkg/storage: Make stacktraceKey []byte and use inlined map[string]