--- a/pkg/storage/series_test.go
+++ b/pkg/storage/series_test.go
@@ -70,8 +70,8 @@ func TestMemSeries(t *testing.T) {
 	require.NoError(t, err)
 
 	require.Len(t, s.samples, 2)
-	require.Equal(t, chunkenc.FromValuesXOR(1), s.samples[k11][0])
-	require.Equal(t, chunkenc.FromValuesXOR(2), s.samples[k12][0])
+	require.Equal(t, chunkenc.FromValuesXOR(1), s.samples[string(k11)][0])
+	require.Equal(t, chunkenc.FromValuesXOR(2), s.samples[string(k12)][0])
 
 	s2 := makeSample(3, []uuid.UUID{uuid2, uuid1})
 	fp2 := &FlatProfile{
@@ -89,8 +89,8 @@ func TestMemSeries(t *testing.T) {
 	require.NoError(t, err)
 
 	require.Len(t, s.samples, 2)
-	require.Equal(t, chunkenc.FromValuesXOR(1, 3), s.samples[k11][0])
-	require.Equal(t, chunkenc.FromValuesXOR(2), s.samples[k12][0]) // sparse - nothing added
+	require.Equal(t, chunkenc.FromValuesXOR(1, 3), s.samples[string(k11)][0])
+	require.Equal(t, chunkenc.FromValuesXOR(2), s.samples[string(k12)][0]) // sparse - nothing added
 
 	// Add another sample with one new Location
 	s3 := makeSample(4, []uuid.UUID{uuid3, uuid1})
@@ -111,9 +111,9 @@ func TestMemSeries(t *testing.T) {
 	require.NoError(t, err)
 
 	require.Len(t, s.samples, 3)
-	require.Equal(t, chunkenc.FromValuesXOR(1, 3), s.samples[k11][0]) // sparse - nothing added
-	require.Equal(t, chunkenc.FromValuesXOR(2), s.samples[k12][0])    // sparse - nothing added
-	require.Equal(t, chunkenc.FromValuesXORAt(2, 4), s.samples[k3][0])
+	require.Equal(t, chunkenc.FromValuesXOR(1, 3), s.samples[string(k11)][0]) // sparse - nothing added
+	require.Equal(t, chunkenc.FromValuesXOR(2), s.samples[string(k12)][0])    // sparse - nothing added
+	require.Equal(t, chunkenc.FromValuesXORAt(2, 4), s.samples[string(k3)][0])
 
 	// Merging another profileTree onto the existing one with one new Location
 	s4 := makeSample(6, []uuid.UUID{uuid5, uuid2, uuid1})
@@ -133,10 +133,10 @@ func TestMemSeries(t *testing.T) {
 	require.NoError(t, err)
 
 	require.Len(t, s.samples, 4)
-	require.Equal(t, chunkenc.FromValuesXOR(1, 3), s.samples[k11][0])  // sparse - nothing added
-	require.Equal(t, chunkenc.FromValuesXOR(2), s.samples[k12][0])     // sparse - nothing added
-	require.Equal(t, chunkenc.FromValuesXORAt(2, 4), s.samples[k3][0]) // sparse - nothing added
-	require.Equal(t, chunkenc.FromValuesXORAt(3, 6), s.samples[k4][0])
+	require.Equal(t, chunkenc.FromValuesXOR(1, 3), s.samples[string(k11)][0])  // sparse - nothing added
+	require.Equal(t, chunkenc.FromValuesXOR(2), s.samples[string(k12)][0])     // sparse - nothing added
+	require.Equal(t, chunkenc.FromValuesXORAt(2, 4), s.samples[string(k3)][0]) // sparse - nothing added
+	require.Equal(t, chunkenc.FromValuesXORAt(3, 6), s.samples[string(k4)][0])
 
 	// Merging another profileTree onto the existing one with one new Location
 	s5 := makeSample(7, []uuid.UUID{uuid2, uuid1})
@@ -154,10 +154,10 @@ func TestMemSeries(t *testing.T) {
 	require.NoError(t, err)
 
 	require.Len(t, s.samples, 4)
-	require.Equal(t, chunkenc.FromValuesXOR(1, 3, 0, 0, 7), s.samples[k11][0])
-	require.Equal(t, chunkenc.FromValuesXOR(2), s.samples[k12][0])     // sparse - nothing added
-	require.Equal(t, chunkenc.FromValuesXORAt(2, 4), s.samples[k3][0]) // sparse - nothing added
-	require.Equal(t, chunkenc.FromValuesXORAt(3, 6), s.samples[k4][0]) // sparse - nothing added
+	require.Equal(t, chunkenc.FromValuesXOR(1, 3, 0, 0, 7), s.samples[string(k11)][0])
+	require.Equal(t, chunkenc.FromValuesXOR(2), s.samples[string(k12)][0])     // sparse - nothing added
+	require.Equal(t, chunkenc.FromValuesXORAt(2, 4), s.samples[string(k3)][0]) // sparse - nothing added
+	require.Equal(t, chunkenc.FromValuesXORAt(3, 6), s.samples[string(k4)][0]) // sparse - nothing added
 
 	require.Equal(t, uint16(5), s.numSamples)
 	require.Equal(t, chunkenc.FromValuesDelta(1000, 2000, 3000, 4000, 5000), s.timestamps[0].chunk)
@@ -206,12 +206,12 @@ func TestMemSeriesMany(t *testing.T) {
 
 	require.Len(t, s.samples, 2)
 
-	it = NewMultiChunkIterator(s.samples[k1])
+	it = NewMultiChunkIterator(s.samples[string(k1)])
 	for i := 1; i < 200; i++ {
 		require.True(t, it.Next())
 		require.Equal(t, int64(i), it.At())
 	}
-	it = NewMultiChunkIterator(s.samples[k2])
+	it = NewMultiChunkIterator(s.samples[string(k2)])
 	for i := 1; i < 200; i++ {
 		require.True(t, it.Next())
 		require.Equal(t, int64(2*i), it.At())