StacktraceKey benchmarks: :tada: 
```
name                  old time/op    new time/op    delta
MakeStacktraceKey-24    2.07µs ± 1%    0.37µs ± 6%   ~     (p=0.100 n=3+3)

name                  old alloc/op   new alloc/op   delta
MakeStacktraceKey-24      625B ± 0%      192B ± 0%   ~     (p=0.100 n=3+3)

name                  old allocs/op  new allocs/op  delta
MakeStacktraceKey-24      22.0 ± 0%       5.0 ± 0%   ~     (p=0.100 n=3+3)
```

DB Appends: :thinking: 
```
name        old time/op    new time/op    delta
Appends-24    16.8ms ± 1%    23.2ms ± 1%   ~     (p=0.100 n=3+3)

name        old alloc/op   new alloc/op   delta
Appends-24    9.84MB ± 0%   11.81MB ± 0%   ~     (p=0.100 n=3+3)

name        old allocs/op  new allocs/op  delta
Appends-24      140k ± 0%      149k ± 0%   ~     (p=0.100 n=3+3)
```
I don't really have a good explanation for this one really. Has to be something where the compiler cannot optimize the `[]byte` to `map[string]...`. I'll follow up with a PR that will only compute the stacktraceKeys once and it'll probably get way better anyway, plus changing to UUIDs for stacktrace IDs soon. 

DB Iterator: :tada: 
```
name         old time/op    new time/op    delta
Iterator-24    61.9µs ± 1%    60.0µs ± 1%   ~     (p=0.100 n=3+3)

name         old alloc/op   new alloc/op   delta
Iterator-24      551B ± 0%      490B ± 0%   ~     (p=0.100 n=3+3)

name         old allocs/op  new allocs/op  delta
Iterator-24      3.00 ± 0%      3.00 ± 0%   ~     (all equal)
```